Fix softlockup in tx_bottom by preventing repeated failed SG fills
When processing transmit packets in tx_bottom, the code was calling
r8152_tx_agg_sg_fill for the same skb even when the SG fill failed to
consume the skb. This could happen if the condition "1 +
skb_shinfo(skb)->nr_frags > RTL_MAX_SG_NUM" was true, leading to
infinite loops and softlockups.

The softlockup here was reproducible only when accessing
huggingface.co with Firefox. With this patch, the warning is
logged as expected and no softlockup occurs anymore:

[212844.954702] r8152 4-1:1.0 enp5s0f4u1: num_sgs 18 > max_sg_num 16

Increasing RTL_MAX_SG_NUM to avoid the failure appears to work too but
it looks more fragile. Furthermore RTL_MAX_SG_NUM is used throughout
the driver to communicate with the network layer and the driver is
meant to be able to consume that number of network buffers at once,
which may not be the case if increasing the value.

This change prevents the infinite loop by invoking the non SG fill to
consume the skb if the SG fill cannot consume it.

Link: https://github.com/wget/realtek-r8152-linux/pull/48
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>